move DealmakingConfiger and SealingConfiger away from config/types.go
