+++ node/config/doc_gen.go
@@ -340,14 +340,6 @@ see https://docs.filecoin.io/mine/lotus/miner-configuration/#using-filters-for-f
 			Comment: ``,
 		},
 	},
-	"DealmakingConfiger interface": []DocField{
-		{
-			Name: "GetDealmakingConfig()",
-			Type: "DealmakingConfig",
-
-			Comment: ``,
-		},
-	},
 	"FeeConfig": []DocField{
 		{
 			Name: "DefaultMaxFee",
@@ -802,14 +794,6 @@ submitting proofs to the chain individually`,
 		{
 			Name: "TerminateBatchWait",
 			Type: "Duration",
-
-			Comment: ``,
-		},
-	},
-	"SealingConfiger interface": []DocField{
-		{
-			Name: "GetSealingConfig()",
-			Type: "SealingConfig",
 
 			Comment: ``,
 		},
