+++ node/config/types.go
@@ -52,32 +52,6 @@ type StorageMiner struct {
 	DAGStore   DAGStoreConfig
 }
 
-type DealmakingConfiger interface {
-	GetDealmakingConfig() DealmakingConfig
-	SetDealmakingConfig(DealmakingConfig)
-}
-
-func (c *StorageMiner) GetDealmakingConfig() DealmakingConfig {
-	return c.Dealmaking
-}
-
-func (c *StorageMiner) SetDealmakingConfig(other DealmakingConfig) {
-	c.Dealmaking = other
-}
-
-type SealingConfiger interface {
-	GetSealingConfig() SealingConfig
-	SetSealingConfig(SealingConfig)
-}
-
-func (c *StorageMiner) GetSealingConfig() SealingConfig {
-	return c.Sealing
-}
-
-func (c *StorageMiner) SetSealingConfig(other SealingConfig) {
-	c.Sealing = other
-}
-
 type DAGStoreConfig struct {
 	// Path to the dagstore root directory. This directory contains three
 	// subdirectories, which can be symlinked to alternative locations if
